# docs/04-api/openapi.yaml
openapi: 3.0.3
info:
  title: ContratosStands API (IIMP)
  version: 0.1.0-draft
  description: >
    API del sistema **ContratosStands** — handler/procesador de reservas de stands
    para eventos IIMP (Perumin, ProExplo, WMC, GESS). BORRADOR derivado del onboarding
    técnico; los contratos con el sistema de facturación (John/SAP) están **por
    confirmar**. Este documento es la fuente de verdad (Swagger/OpenAPI) de la API.
  contact:
    name: Equipo ContratosStands
servers:
  - url: http://localhost:3000/api
    description: Desarrollo local
  - url: https://dev.contratos-stands.iimp/api
    description: Ambiente Dev (PC)
  - url: https://qa.contratos-stands.iimp/api
    description: Ambiente QA (PC)
...
security:
  - bearerAuth: []
paths:
...
  /eventos:
    get:
      tags: [Eventos]
      summary: Lista eventos (versiones) por evento padre
      parameters:
        - name: eventoPadreId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Evento"
...
  /empresas:
    get:
      tags: [Empresas]
      summary: Busca empresas (proxy hacia la fuente única / John / Niel)
      parameters:
        - name: q
          in: query
          description: Razón social o RUC
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Empresa"
...
  /auth/cambiar-password:
    post:
      tags: [Auth]
      summary: Cambia la contraseña temporal (primer ingreso) y reemite la sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [passwordActual, passwordNueva]
              properties:
                passwordActual: { type: string }
                passwordNueva: { type: string, minLength: 8 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [ok, requiereValidarDatos]
                properties:
                  ok: { type: boolean }
                  requiereValidarDatos: { type: boolean, description: "La empresa aún debe validar sus datos" }
        "400":
          description: Contraseña actual incorrecta o nueva inválida
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Evento:
      type: object
      required: [id, eventoPadreId, tipoEvento, codigoEvento, estado]
      properties:
        id: { type: string }
        eventoPadreId: { type: string }
        tipoEvento: { type: integer, example: 14 }
        codigoEvento: { type: integer, example: 1 }
        planoRef: { type: string }
        anio: { type: string, example: "2026" }
        estado:
          type: string
          enum: [draft, active, closed, cancelled]
...
    Empresa:
      type: object
      properties:
        id: { type: string }
        razonSocial: { type: string }
        ruc: { type: string }
        origen:
          type: string
          enum: [john, niel, local]
...